moved wrap_litserve_start to utils
