+++ src/litserve/utils.py
@@ -16,6 +16,8 @@ import logging
 import pickle
 import uuid
 from typing import Coroutine, Optional
+from contextlib import contextmanager
+from litserve.server import LitServer
 
 from fastapi import HTTPException
 from starlette.middleware.base import BaseHTTPMiddleware
@@ -68,6 +70,16 @@ async def azip(*async_iterables):
             break
         yield tuple(results)
 
+@contextmanager
+def wrap_litserve_start(server: LitServer):
+    server.app.response_queue_id = 0
+    if server.lit_spec:
+        server.lit_spec.response_queue_id = 0
+    manager, processes = server.launch_inference_worker(num_uvicorn_servers=1)
+    yield server
+    for p in processes:
+        p.terminate()
+    manager.shutdown()
 
 class MaxSizeMiddleware(BaseHTTPMiddleware):
     def __init__(
