+++ tests/test_auth.py
@@ -18,7 +18,7 @@ from fastapi.testclient import TestClient
 
 from litserve import LitAPI, LitServer
 import litserve.server
-from tests.conftest import wrap_litserve_start
+from litserve.utils import wrap_litserve_start
 
 
 class SimpleAuthedLitAPI(LitAPI):
