+++ tests/test_compression.py
@@ -14,7 +14,7 @@
 
 from fastapi import Request, Response
 from fastapi.testclient import TestClient
-from tests.conftest import wrap_litserve_start
+from litserve.utils import wrap_litserve_start
 from litserve import LitAPI, LitServer
 
 # trivially compressible content
